feat: add validation, authorization logic and middleware integration
- Implemented validation rules for input models
- Added authorization checks across protected endpoints
- Integrated middleware for token and user identity handling
- Successfully tested end-to-end front-back operations

+++ backend/BillingSystemOperational/BillingSystemOperational.InvoiceService/WebApi/Controllers/InvoiceController.cs
@@ -1,8 +1,10 @@
 ﻿using BillingSystemOperational.InvoiceService.Application.Dto;
 using BillingSystemOperational.InvoiceService.Application.Interface;
+using FluentValidation;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.IdentityModel.Tokens;
 
 namespace BillingSystemOperational.InvoiceService.WebApi.Controllers
 {
@@ -11,10 +13,27 @@ namespace BillingSystemOperational.InvoiceService.WebApi.Controllers
     public class InvoiceController : ControllerBase
     {
         private readonly IInvoiceService _invoiceService;
-
-        public InvoiceController(IInvoiceService invoiceService)
+        private readonly IValidator<InvoiceUpdateDto> _updateValidator;
+        private readonly IValidator<InvoiceSaveDto> _saveValidator;
+        private readonly IValidator<InvoiceDeleteRequestDto> _deleteValidator;
+        private readonly IValidator<InvoiceLineSaveDto> _invoiceLineValidator;
+        private readonly IValidator<InvoiceListItemRequestDto> _invoiceListValidator;
+        private readonly IValidator<InvoiceLineUpdateDto> _invoiceLineUpdateValidator;
+        public InvoiceController(IInvoiceService invoiceService,
+            IValidator<InvoiceUpdateDto> updateValidator,
+            IValidator<InvoiceSaveDto> saveValidator,
+            IValidator<InvoiceDeleteRequestDto> deleteValidator,
+            IValidator<InvoiceLineSaveDto> invoiceLineValidator,
+            IValidator<InvoiceListItemRequestDto> invoiceListValidator,
+            IValidator<InvoiceLineUpdateDto> invoiceLineUpdateValidator)
         {
             _invoiceService = invoiceService;
+            _updateValidator = updateValidator;
+            _saveValidator = saveValidator;
+            _deleteValidator = deleteValidator;
+            _invoiceLineValidator = invoiceLineValidator;
+            _invoiceListValidator = invoiceListValidator;
+            _invoiceLineUpdateValidator = invoiceLineUpdateValidator;
         }
 
         [Authorize]
@@ -23,6 +42,7 @@ namespace BillingSystemOperational.InvoiceService.WebApi.Controllers
         {
             return Ok(HttpContext.Request.Headers.Authorization.ToString());
         }
+        [Authorize]
         [HttpGet("getivoices")]
         public async Task<IActionResult> GetListInvoice()
         {
@@ -33,9 +53,14 @@ namespace BillingSystemOperational.InvoiceService.WebApi.Controllers
             }
             return Ok(result);
         }
+        [Authorize]
         [HttpGet]
         public async Task<IActionResult> GetInvoice(int id)
         {
+            if (id.ToString().IsNullOrEmpty())
+            {
+                return BadRequest("id value is null");
+            }
             var result = await _invoiceService.GetInvoice(id);
             if (!result.Success)
             {
@@ -43,9 +68,15 @@ namespace BillingSystemOperational.InvoiceService.WebApi.Controllers
             }
             return Ok(result.Data);
         }
+        [Authorize]
         [HttpGet("getlist")]
-        public async Task<IActionResult> InvoiceList(InvoiceListItemRequestDto requestDto)
+        public async Task<IActionResult> InvoiceList([FromQuery] InvoiceListItemRequestDto requestDto)
         {
+            var validate = await _invoiceListValidator.ValidateAsync(requestDto);
+            if (!validate.IsValid)
+            {
+                return BadRequest(validate.Errors);
+            }
             var result = await _invoiceService.GetInvoiceList(requestDto);
             if (!result.Success)
             {
@@ -53,16 +84,28 @@ namespace BillingSystemOperational.InvoiceService.WebApi.Controllers
             }
             return Ok(result);
         }
+        [Authorize]
         [HttpPost]
         public async Task<IActionResult> InvoiceSave(InvoiceSaveDto request)
         {
+            var validate = await _saveValidator.ValidateAsync(request);
+            if (!validate.IsValid)
+            {
+                return BadRequest(validate.Errors);
+            }
             var result = await _invoiceService.Add(request);
             if (!result.Success) { return BadRequest(result); }
             return Ok(result);
         }
+        [Authorize]
         [HttpPut]
         public async Task<IActionResult> InvoiceUpdate(InvoiceUpdateDto request)
         {
+            var validate = await _updateValidator.ValidateAsync(request);
+            if (!validate.IsValid)
+            {
+                return BadRequest(validate.Errors);
+            }
             var result = await _invoiceService.Update(request);
             if (!result.Success)
             {
@@ -70,9 +113,15 @@ namespace BillingSystemOperational.InvoiceService.WebApi.Controllers
             }
             return Ok(result);
         }
+        [Authorize]
         [HttpDelete]
         public async Task<IActionResult> InvoiceDelete(InvoiceDeleteRequestDto request)
         {
+            var validate = await _deleteValidator.ValidateAsync(request);
+            if (!validate.IsValid)
+            {
+                return BadRequest(validate.Errors);
+            }
             var result = await _invoiceService.Delete(request);
             if (!result.Success)
             {
